serdect::slice

Function deserialize_hex_or_bin

pub fn deserialize_hex_or_bin<'de, D>(
    buffer: &mut [u8],
    deserializer: D,
) -> Result<&[u8], D::Error>
where
    D: Deserializer<'de>,

Deserialize from hex when using human-readable formats or binary if the format is binary. Fails if the buffer is smaller then the resulting slice.