serdect::slice

Function serialize_hex_lower_or_bin

pub fn serialize_hex_lower_or_bin<S, T>(
    value: &T,
    serializer: S,
) -> Result<S::Ok, S::Error>
where
    S: Serializer,
    T: AsRef<[u8]>,

Serialize the given type as lower case hex when using human-readable formats or binary if the format is binary.